requirements.txt
certifi==2020.6.20
cffi==1.14.3
cryptography==3.2.1
echo==0.5
future==0.18.2
numpy==1.19.3
pycparser==2.20
python-telegram-bot==11.1.0
QtPy==1.9.0
requests-debugger==0.0.3
six==1.15.0
